Major refactoring using new generation capabilities provided by sundrio.
Istio resources are now automatically generated and entities corresponding to Istio CRDs (e.g. `Gateway`) are now directly accessible when it was previously needed to go through `IstioResource` via a `spec` object. To keep the `Gateway` example, it was previously needed to create one via an `IstioResourceBuilder` and the `Gateway` object was the `spec` that Istio defined. Now, a `Gateway` is an Istio resource that can be built using a `GatewayBuilder` and is associated to a `GatewaySpec` which better reflects how we think of these entities. The client API also now more closely follows the `kubernetes-client` API with better grouping of builder methods to allow for cleaner interfaces. Thanks a lot to @iocanel for this massive work!