Fix handling of plurals
With the previous implementation plurals (which determine the API path
that the Kubernetes Client uses) where handled by simply appending an
"s" to the name of the class. This fails for a number of CRD, like
Policy.
We now use the information that is present in the crdName
during the Operations generation