fix: use detyped map for Struct in EnvoyFilter
Tried to devise a type-safe solution but didn't figure out a quick way
to do it. Considering that EnvoyFilters are not used anymore in more
recent istio versions, not going to spend more efforts on this so went
with the detyped solution suggested by @zhengzepeng
in https://github.com/snowdrop/istio-java-api/issues/99#issuecomment-653317222
Thanks for the investigation!

Fixes #105
Fixes #99